class WorkflowAdapterImplementation
